+++ src/app/src/screens/Network/CreateDrawer.tsx
@@ -1,111 +1,153 @@
-import { useState, memo } from "react";
+import { useState, useMemo } from "react";
 import {
   ButtonGroup,
   Button,
   Intent,
-  InputGroup,
-  FormGroup,
   Drawer,
   DrawerSize,
   ProgressBar,
   Classes,
-  TextArea
 } from "@blueprintjs/core";
 import { IconNames } from "@blueprintjs/icons";
-import { useForm, Controller } from "react-hook-form";
+import { useForm, FormProvider, useFormContext } from "react-hook-form";
 import { useTranslation } from "react-i18next";
-import isEqual from "react-fast-compare";
 
 // project
 import { useStoreActions } from "../../domain/types";
+import { createNetworkSubnet, NetworkSubnetItem, NetworkSubnetsForm } from "./NetworkSubnetsForm";
+import dayjs from "dayjs";
+import { NetworkSubnet } from "../../Types.container-app";
+import { Notification } from "../../Notification";
+import { NetworkPropertiesForm } from "./NetworkPropertiesForm";
 
-// Secret drawer
+// Drawer
+
+export interface CreateFormData {
+  networkName: "",
+  networkInterface: string;
+  dnsEnabled: boolean;
+  internal: boolean;
+  ipv6Enabled: boolean;
+  driver: string;
+  subnets: NetworkSubnetItem[];
+}
+
+export function toNetworkSubnets(subnets: NetworkSubnetItem[]) {
+  const items: NetworkSubnet[] = subnets.map((it) => {
+    const coerced = {
+      ...it,
+      guid: undefined
+    };
+    return coerced;
+  }).reduce<NetworkSubnet[]>((acc, it) => {
+    if (it.gateway && it.subnet) {
+      acc.push(it);
+    }
+    return acc;
+  }, []);
+  return items;
+}
+
+export interface FormActionsProps {
+  pending?: boolean;
+}
+export const FormActions: React.FC<FormActionsProps> = ({ pending }) => {
+  const { t } = useTranslation();
+  const { formState } = useFormContext();
+  const pendingIndicator = (
+    <div className="AppDrawerPendingIndicator">{pending && <ProgressBar intent={Intent.SUCCESS} />}</div>
+  );
+  return (
+    <>
+      <ButtonGroup fill>
+        <Button
+          disabled={pending || !formState.isValid}
+          intent={Intent.PRIMARY}
+          icon={IconNames.GRAPH}
+          title={t("Click to launch creation")}
+          text={t("Create")}
+          type="submit"
+        />
+      </ButtonGroup>
+      {pendingIndicator}
+    </>
+  );
+}
+
 export interface CreateDrawerProps {
   onClose: () => void;
 }
-export const CreateDrawer: React.FC<CreateDrawerProps> = memo(
-  ({ onClose }) => {
-    const { t } = useTranslation();
-    const { control, handleSubmit } = useForm<{
-      networkName: string;
-    }>();
-    const [pending, setPending] = useState(false);
-    const networkCreate = useStoreActions((actions) => actions.network.networkCreate);
-    const onSubmit = handleSubmit(async (data) => {
-      try {
-        setPending(true);
-        // await networkCreate({
-        //   name: data.networkName,
-        // });
-        onClose();
-      } catch (error) {
-        console.error("Unable to create network", error);
-      } finally {
-        setPending(false);
-      }
-    });
-    const pendingIndicator = (
-      <div className="AppDrawerPendingIndicator">{pending && <ProgressBar intent={Intent.SUCCESS} />}</div>
-    );
-    return (
-      <Drawer
-        className="AppDrawer"
-        icon={IconNames.PLUS}
-        title={t("Create new network")}
-        usePortal
-        size={DrawerSize.SMALL}
-        onClose={onClose}
-        isOpen
-        hasBackdrop={false}
-      >
-        <div className={Classes.DRAWER_BODY}>
-          <form className={Classes.DIALOG_BODY} onSubmit={onSubmit}>
-            <ButtonGroup fill>
-              <Button
-                disabled={pending}
-                intent={Intent.PRIMARY}
-                icon={IconNames.KEY}
-                title={t("Click to launch creation")}
-                text={t("Create")}
-                type="submit"
-              />
-            </ButtonGroup>
-            {pendingIndicator}
+export const CreateDrawer: React.FC<CreateDrawerProps> = ({ onClose }) => {
+  const { t } = useTranslation();
+  const subnets = useMemo(() => {
+    return [createNetworkSubnet()];
+  }, []);
+  const methods = useForm<CreateFormData>({
+    mode: "all",
+    reValidateMode: "onChange",
+    shouldUseNativeValidation: false,
+    defaultValues: {
+      networkName: "",
+      networkInterface: "",
+      dnsEnabled: false,
+      internal: false,
+      ipv6Enabled: false,
+      driver: "",
+      subnets,
+    }
+  });
+  const { handleSubmit } = methods;
+  const [pending, setPending] = useState(false);  // Form initial data
+
+  const networkCreate = useStoreActions((actions) => actions.network.networkCreate);
+  const onSubmit = handleSubmit(async (data) => {
+    setPending(true);
+    try {
+      await networkCreate({
+        created: dayjs().toISOString(),
+        dns_enabled: data.dnsEnabled,
+        driver: data.driver,
+        internal: data.internal,
+        name: data.networkName,
+        network_interface: data.networkInterface,
+        subnets: toNetworkSubnets(data.subnets),
+      });
+      onClose();
+      Notification.show({ message: t("Network has been created"), intent: Intent.SUCCESS });
+    } catch (error: any) {
+      Notification.show({
+        message: t("{{message}} - {{data}}", {
+          message: error.message || t("Command failed"),
+          data: error.details?.result?.result?.data?.cause,
+        }),
+        intent: Intent.DANGER
+      });
+    } finally {
+      setPending(false);
+    }
+  });
+  return (
+    <Drawer
+      className="AppDrawer"
+      icon={IconNames.PLUS}
+      title={t("Create new network")}
+      usePortal
+      size={DrawerSize.SMALL}
+      onClose={onClose}
+      isOpen
+      hasBackdrop={false}
+    >
+      <div className={Classes.DRAWER_BODY}>
+        <FormProvider {...methods}>
+          <form name="CreateNetworkForm" className={Classes.DIALOG_BODY} onSubmit={onSubmit}>
+            <FormActions />
             <div className="AppDataForm" data-form="network.create">
-              <FormGroup disabled={pending} label={t("Name")} labelFor="networkName" labelInfo="(required)">
-                <Controller
-                  control={control}
-                  name="networkName"
-                  rules={{ required: true }}
-                  defaultValue=""
-                  render={({ field: { onChange, onBlur, value, name, ref }, fieldState: { invalid }, formState }) => {
-                    return (
-                      <InputGroup
-                        fill
-                        autoFocus
-                        disabled={pending || formState.isSubmitting || formState.isValidating}
-                        id={name}
-                        className="networkName"
-                        placeholder={t("Type to set a name")}
-                        name={name}
-                        value={value}
-                        required
-                        onBlur={onBlur}
-                        onChange={onChange}
-                        inputRef={ref}
-                        intent={invalid ? Intent.DANGER : Intent.NONE}
-                      />
-                    );
-                  }}
-                />
-              </FormGroup>
+              <NetworkPropertiesForm disabled={pending} />
+              <NetworkSubnetsForm subnets={subnets} disabled={pending} />
             </div>
           </form>
-        </div>
-      </Drawer>
-    );
-  },
-  (prev, next) => {
-    return isEqual(prev, next);
-  }
-);
+        </FormProvider>
+      </div>
+    </Drawer>
+  );
+}